+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `Authorization` module with Casbin
 - Add `EnforceScope` type and module
 - Add `Session.Authorize.withScope` function
+- Add `JWT.TokenData` field `Subject`
 
 ## 8.0.0 - 2026-01-14
 - [**BC**] Rename